Move some generic stuff from xcb.h to display.h
