+++ test/box-test.c
@@ -39,6 +39,7 @@
 #include "rofi.h"
 #include "xrmoptions.h"
 #include "helper.h"
+#include "display.h"
 unsigned int test =0;
 #define TASSERT( a )    {                                 \
         assert ( a );                                     \
